Refactor imports and update dependencies
Updated README.md example to use new MattermostClient login method.
Updated package references in Mattermost.Tests.csproj and Mattermost.csproj.
Refactored imports and namespaces in various files for better organization.
Removed handling of the Unknown event in MattermostClient.cs.
