+++ Sources/Mattermost/Mattermost.csproj
@@ -28,7 +28,7 @@
 		<None Include="..\..\README.md" Pack="true" PackagePath="\" />
 		<None Include="package-icon.png" Pack="true" PackagePath="\" />
 		<None Include="..\..\LICENSE.md" Pack="true" PackagePath="LICENSE.md" />
-		<PackageReference Include="System.Text.Json" Version="8.0.4" />
+		<PackageReference Include="System.Text.Json" Version="9.0.1" />
 		<PackageReference Include="Microsoft.SourceLink.GitHub" Version="8.0.0" PrivateAssets="All" />
 	</ItemGroup>
 </Project>
